Remove wxt/browser/webextension-polyfill
