+++ packages/wxt/typedoc.json
@@ -5,7 +5,6 @@
     "src/sandbox/index.ts",
     "src/browser/index.ts",
     "src/browser/chrome.ts",
-    "src/browser/webextension-polyfill.ts",
     "src/index.ts",
     "src/storage.ts",
     "src/modules.ts"
